sdhci: introduce ops->pre_voltage_switch and extend voltage_switch
Provide a hook to be called *before* a voltage switch takes place. This
may be required, for example, to configure pins as 3.3V tolerant prior
to switching back from 1.8V to 3.3V.

This adds a 'struct mmc_ios' arg to voltage_switch to signal which
voltage we have just switched to.